Refactor imports to allow for OAuth with imports, move issue providers to projects django app
