+++ rdmo/projects/mixins.py
@@ -1,9 +1,10 @@
 from collections import defaultdict
+from pathlib import Path
 
 from django.contrib.sites.shortcuts import get_current_site
 from django.core.exceptions import ValidationError
 from django.http import HttpResponseRedirect, Http404, HttpResponse
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import get_object_or_404, render, redirect
 from django.utils.translation import gettext_lazy as _
 
 from rdmo.core.imports import handle_uploaded_file, file_path_exists
@@ -71,83 +72,36 @@ class ProjectImportMixin(object):
 
         return import_plugin
 
-    def render(self):
-        import_key = self.kwargs.get('format')
-        current_project = self.object
-
-        try:
-            return self.get_import_plugin(import_key, current_project).render()
-        except (NotImplementedError, Http404):
-            return HttpResponseRedirect(self.success_url)
-
-    def submit(self):
-        import_key = self.kwargs.get('format')
-        current_project = self.object
-
-        try:
-            import_plugin = self.get_import_plugin(import_key, current_project)
-        except (NotImplementedError, Http404):
-            return HttpResponseRedirect(self.success_url)
-
-        response = import_plugin.submit()
-
-        if isinstance(response, HttpResponse):
-            # this is a django response, return it!
-            return response
-
-        elif isinstance(response, str) and file_path_exists(response):
-            import_plugin.file_name = response
-
-            if import_plugin.check():
-                # this is a valid file path
-                try:
-                    import_plugin.process()
-                except ValidationError as e:
-                    return render(self.request, 'core/error.html', {
-                        'title': _('Import error'),
-                        'errors': e
-                    }, status=400)
-
-                # store information in session for ProjectCreateImportView
-                self.request.session['update_import_tmpfile_name'] = import_plugin.file_name
-                self.request.session['update_import_key'] = import_key
-
-                # attach questions and current values
-                self.update_values(current_project, import_plugin.catalog, import_plugin.values, import_plugin.snapshots)
-
-                return render(self.request, 'projects/project_import.html', {
-                    'method': 'import_file',
-                    'current_project': current_project,
-                    'source_title': import_plugin.source_title,
-                    'source_project': import_plugin.project,
-                    'values': import_plugin.values,
-                    'snapshots': import_plugin.snapshots if not current_project else None,
-                    'tasks': import_plugin.tasks,
-                    'views': import_plugin.views
-                })
-            else:
-                return render(self.request, 'core/error.html', {
-                    'title': _('Import error'),
-                    'errors': [_('Files of this type cannot be imported.')]
-                }, status=400)
-
-        else:
-            return None
-
     def upload_file(self):
-        current_project = self.object
-
         try:
             uploaded_file = self.request.FILES['uploaded_file']
         except KeyError:
-            return None
+            return render(self.request, 'core/error.html', {
+                'title': _('Import error'),
+                'errors': [_('There has been an error with your import.')]
+            }, status=400)
         else:
-            import_tmpfile_name = handle_uploaded_file(uploaded_file)
+            self.request.session['import_file_name'] = handle_uploaded_file(uploaded_file)
+            self.request.session['import_source_title'] = uploaded_file.name
+
+            # redirect to the import form
+            return redirect(self.request.path)
+
+    def import_form(self):
+        current_project = self.object
+
+        import_file_name = self.request.session.get('import_file_name')
+        import_source_title = self.request.session.get('import_source_title')
+        if import_file_name is None or not Path(import_file_name).is_file():
+            return render(self.request, 'core/error.html', {
+                'title': _('Import error'),
+                'errors': [_('There has been an error with your import. No uploaded or retrieved file could be found.')]
+            }, status=400)
 
         for import_key, import_plugin in get_plugins('PROJECT_IMPORTS').items():
-            import_plugin.source_title = uploaded_file.name
-            import_plugin.file_name = import_tmpfile_name
             import_plugin.current_project = current_project
+            import_plugin.file_name = import_file_name
+            import_plugin.source_title = import_source_title
 
             if import_plugin.upload and import_plugin.check():
                 try:
@@ -159,8 +113,7 @@ class ProjectImportMixin(object):
                     }, status=400)
 
                 # store information in session for ProjectCreateImportView
-                self.request.session['update_import_tmpfile_name'] = import_tmpfile_name
-                self.request.session['update_import_key'] = import_key
+                self.request.session['import_key'] = import_key
 
                 # attach questions and current values
                 self.update_values(current_project, import_plugin.catalog, import_plugin.values, import_plugin.snapshots)
@@ -184,8 +137,8 @@ class ProjectImportMixin(object):
     def import_file(self):
         current_project = self.object
 
-        import_tmpfile_name = self.request.session.get('update_import_tmpfile_name')
-        import_key = self.request.session.get('update_import_key')
+        import_tmpfile_name = self.request.session.pop('import_file_name')
+        import_key = self.request.session.pop('import_key')
         checked = [key for key, value in self.request.POST.items() if 'on' in value]
 
         if import_tmpfile_name and import_key:
@@ -229,6 +182,11 @@ class ProjectImportMixin(object):
                     save_import_views(import_plugin.project, import_plugin.views)
 
                     return HttpResponseRedirect(import_plugin.project.get_absolute_url())
+
+        return render(self.request, 'core/error.html', {
+            'title': _('Import error'),
+            'errors': [_('There has been an error with your import.')]
+        }, status=400)
 
     def import_project(self):
         current_project = self.object
